Upgrade to mainline-tracking-pre-prod-v6.17-rc3-linux-250910T070627Z
Integreate the Intel Linux mainline-tracking kernel into the EMT to
support the platform which are not ready in Linux long-term support kernel.

Furthermore, we also have below changes based on v6.12 kernel:

1.Add cpupower-service.conf into kernel-tools package in kernel.spec

Since 6.15, cpupower-service.conf is presented and it need be added
into the kernel-tools package.

e5174365c1324 cpupower: do not install files to /etc/default/

2. Remove the perf-core directory from kernel-tools package in kernel.spec

Since 6.13, the perf-core/* are removed.

3cc550f5bbcf5 perf tools: Remove dependency on libaudit

Signed-off-by: Baoli.Zhang <baoli.zhang@intel.com>